+++ src/OpenSimBindings/MuscleDecorationStyle.cpp
@@ -10,14 +10,14 @@
 static constexpr auto c_Styles = osc::MakeSizedArray<osc::MuscleDecorationStyle, static_cast<size_t>(osc::MuscleDecorationStyle::TOTAL)>
 (
     osc::MuscleDecorationStyle::OpenSim,
-    osc::MuscleDecorationStyle::Scone,
+    osc::MuscleDecorationStyle::FibersAndTendons,
     osc::MuscleDecorationStyle::Hidden
 );
 
 static constexpr auto c_StyleStrings = osc::MakeSizedArray<char const*, static_cast<size_t>(osc::MuscleDecorationStyle::TOTAL)>
 (
-    "OpenSim (lines)",
-    "SCONE (tendons + fibers)",
+    "OpenSim",
+    "Fibers & Tendons",
     "Hidden"
 );